AB#4786 refactor moving code to a more appropriate location
